+++ FCARDIO.Protocol.Door/Door8800/Card/06DeleteCard/DeleteCardBase.cs
@@ -92,7 +92,7 @@ namespace DoNetDrive.Protocol.Door.Door8800.Card
         /// </summary>
         /// <param name="CardList">错误卡列表</param>
         /// <param name="buf"></param>
-        protected override void ReadCardByFailBuf(List<UInt64> CardList, IByteBuffer buf)
+        protected override void ReadCardByFailBuf(List<decimal> CardList, IByteBuffer buf)
         {
             return;
         }
